+++ chart/irs/values-dev.yaml
@@ -9,3 +9,8 @@ keycloak:
     clientSecret: <path:product-traceability-irs/data/dev/keycloak/oauth2#clientSecret>
     clientTokenUri: <path:product-traceability-irs/data/dev/keycloak/oauth2#tokenUri>
     jwkSetUri: <path:product-traceability-irs/data/dev/keycloak/oauth2#jwkSetUri>
+
+aasProxy:
+  submodel:
+    username: <path:product-traceability-irs/data/dev/aasProxy/submodel#username>
+    password: <path:product-traceability-irs/data/dev/aasProxy/submodel#password>
