+++ chart/irs/values-int.yaml
@@ -21,3 +21,8 @@ keycloak:
     clientSecret: <path:product-traceability-irs/data/int/keycloak/oauth2#clientSecret>
     clientTokenUri: <path:product-traceability-irs/data/int/keycloak/oauth2#tokenUri>
     jwkSetUri: <path:product-traceability-irs/data/int/keycloak/oauth2#jwkSetUri>
+
+aasProxy:
+  submodel:
+    username: <path:product-traceability-irs/data/int/aasProxy/submodel#username>
+    password: <path:product-traceability-irs/data/int/aasProxy/submodel#password>
